requirements.txt
Babel==2.11.0
certifi==2023.11.17
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
ftfy==6.0.3
google-measurement-protocol==1.1.0
idna==3.6
importlib-resources==5.4.0
joblib==1.1.1
jsonpatch==1.32
jsonpointer==2.3
keopscore==2.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
Pillow==8.4.0
prices==1.1.1
pybind11==2.11.1
pykeops==2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
torch==1.10.1
torchnet==0.0.4
torchvision==0.11.2
tornado==6.1
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.18
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.3.1
zipp==3.6.0
